Cập nhập giảm giá, thêm đa ngôn ngữ, thêm chức năng trả hàng, tối ưu code, tối ưu css, cập nhập admin

+++ src/main/java/vn/urbansteps/repository/AdminActionLogRepository.java
@@ -3,6 +3,6 @@ package vn.urbansteps.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.urbansteps.model.AdminActionLog;
 
-public interface AdminActionLogRepository extends JpaRepository<AdminActionLog, Long> {
+public interface AdminActionLogRepository extends JpaRepository<AdminActionLog, Integer> {
     // Có thể thêm các phương thức truy vấn tuỳ ý
 }
